# [improvement] Tracer spans are inexpensive for unsampled operations