--- a/tracing-benchmarks/src/jmh/java/com/palantir/tracing/TracingBenchmark.java
+++ b/tracing-benchmarks/src/jmh/java/com/palantir/tracing/TracingBenchmark.java
@@ -99,7 +99,7 @@ public class TracingBenchmark {
     private static Runnable wrapWithSpan(Runnable toBenNested) {
         return () -> {
             try {
-                Tracer.startSpan("span");
+                Tracer.fastStartSpan("span");
                 toBenNested.run();
             } finally {
                 Tracer.fastCompleteSpan();