--- a/tracing-jersey/src/main/java/com/palantir/tracing/jersey/TraceEnrichingFilter.java
+++ b/tracing-jersey/src/main/java/com/palantir/tracing/jersey/TraceEnrichingFilter.java
@@ -75,7 +75,7 @@ public final class TraceEnrichingFilter implements ContainerRequestFilter, Conta
                 Tracer.fastStartSpan(operation, SpanType.SERVER_INCOMING);
             } else {
                 // caller's span is this span's parent.
-                Tracer.startSpan(operation, spanId, SpanType.SERVER_INCOMING);
+                Tracer.fastStartSpan(operation, spanId, SpanType.SERVER_INCOMING);
             }
         }