--- a/tracing/src/main/java/com/palantir/tracing/DeferredTracer.java
+++ b/tracing/src/main/java/com/palantir/tracing/DeferredTracer.java
@@ -104,7 +104,7 @@ public final class DeferredTracer implements Serializable {
 
         Optional<Trace> originalTrace = Tracer.copyTrace();
 
-        Tracer.setTrace(new Trace(isObservable, traceId));
+        Tracer.setTrace(Trace.of(isObservable, traceId));
         if (parentSpanId != null) {
             Tracer.startSpan(operation, parentSpanId, SpanType.LOCAL);
         } else {