… tracing-jaxrs/src/test/java/com/palantir/tracing/jaxrs/JaxRsTracersTest.java
… tracing-servlet/src/test/java/com/palantir/tracing/servlet/LeakedTraceFilterTest.java
… tracing/src/main/java/com/palantir/tracing/Tracers.java
… tracing/src/test/java/com/palantir/tracing/AsyncSlf4jSpanObserverTest.java
… tracing/src/test/java/com/palantir/tracing/AsyncTracerTest.java
… tracing/src/test/java/com/palantir/tracing/CloseableTracerTest.java
… tracing/src/test/java/com/palantir/tracing/TraceTest.java
… tracing/src/test/java/com/palantir/tracing/TracerTest.java
… tracing/src/test/java/com/palantir/tracing/TracersTest.java