@carterkozak 2019-06-20
Still WIP, some refactors I need to make to reduce code duplication.