@carterkozak 2019-06-21
> 4x speedup and 4x reduction in allocations for common UNDECIDED case

This sounded lower than expected, the benchmark was set to 10% sample rate instead of 1% for the `UNDECIDED` case. I have updated the main comment with the more accurate numbers.